Started APITesters (#359)
For [CSDK-143].

This relies on TypeScript compilation to detect type issues. A separate `tsconfig.json` inherits from the parent one to compile only these files.

`noUnusedLocals` and `noUnusedParameters` are set to `false` to ignore errors due to all these values being unused in the testers.

This PR only adds the main setup. See #360 for the rest of testers.

[CSDK-143]: https://revenuecats.atlassian.net/browse/CSDK-143?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ