Create react-native-purchases-ui (#807)
Creates a sub library called `react-native-purchases-ui`

This library differs in some stuff from `react-native-purchases` and has
been initialized using `create-react-native-library` which uses
`react-native@0.73.1` and a tool `react-native-builder-bob` to prepare
the package. This tool wraps both typescript and babel and generates the
modules and types properly for the library. I picked the default
`esnext` module (we use `commonjs` in the root library) not sure if that
would be a problem, but that's the default and haven't seen any issue.

I tried doing the same we do in `react-native-purchases` (no
`react-native-builder-bob` and older react native version) and had
issues refreshing the javascript sources of the library when modifying
them. With the current setup everything works smootly. After running the
tester app using `yarn example android|ios` the TS sources in either
`react-native-purchases` or `react-native-purchases-ui` are updated
after modification, no need to rerun.

I also added a `.yarn` folder to the repo so everyone uses the latest
yarn available. I was having some issues with `tsc` being run when
running `yarn install` using yarn v1, which is the version I have in my
machine. This version won't run `prepare` on the library unless
indicated to do so. This was also the default for the barebones library
created with `create-react-native-library`.

For now, I removed the yarn caches in the CircleCI config, because I was
having issues with them. We can come back to that later.

There's no code for publishing the library yet. That will come in a
different PR as well.

---------

Co-authored-by: Toni Rico <toni.rico.diez@revenuecat.com>