Use ReactContext instead of ThemedReactContext in PaywallFooterViewManager (#971)
`ThemedReactContext` is broken in the new architecture. This fix
suggested by @cortinico prevents a crash in the new architecture. The
behavior is the same

The PaywallFooter is still broken and
`themedReactContext.getNativeModule(UIManagerModule::class.java)` is
throwing

```
com.facebook.react.bridge.ReactNoCrashBridgeNotAllowedSoftException: getNativeModule(UIManagerModule.class) cannot be called when the bridge is disabled
```