docs: Update installation instructions for Expo (#1041)
- [x] A description about what and why you are contributing, even if
it's trivial.

The instructions under the "Installation" section mentioned steps to be
taken when a project uses ExpoKit. ExpoKit has been deprecated and
removed from the Expo ecosystem, thus, making these instructions
outdated.

The new instructions only require creating a development build when
using the Expo framework in React Native projects. I've added links to
creating a project and also for creating a development build that walks
through all the essential steps required to create a development build.
Also, added the section for Bare workflow (if any of the project uses it
and might want to use Expo inside, the docs links mentioned in this
section will help them).

After a project is created with Expo, only the library's installation
command is required to run.

- [ ] The issue number(s) or PR number(s) in the description if you are
contributing in response to those.

  - [ ] If applicable, unit tests.

---------

Co-authored-by: Keith Kurak <keith.kurak@gmail.com>