Workaround to remove `NSNull` objects in ObjC layer (#1196)
This PR implements a workaround that fixes #1185.

With RN v0.77.1 and 0.78.0, the Objective-C layer of
react-native-purchases is receiving `NSNull` objects for `null`
TypeScript values (also reported in in the [RN
repo](https://github.com/facebook/react-native/pull/49250#issuecomment-2668465893)).

Note that we cannot avoid the console warnings of type 
> JSON value '<null>' of type NSNull cannot be converted to NSString

Keep reading for an explanation.

## Why this was not happening in `DEBUG`

The `RCT_JSON_CONVERTER` macro is defined like this in `RCTConvert.mm`
in the `React-core` dependency:

```
#if RCT_DEBUG
#define RCT_JSON_CONVERTER(type)             \
  +(type *)type : (id)json                   \
  {                                          \
    if ([json isKindOfClass:[type class]]) { \
      return json;                           \
    } else if (json) {                       \
      RCTLogConvertError(json, @ #type);     \
    }                                        \
    return nil;                              \
  }
#else
#define RCT_JSON_CONVERTER(type) \
  +(type *)type : (id)json       \
  {                              \
    return json;                 \
  }
#endif
```

In case of an unrecognized JSON object (e.g. `NSNull`), in `DEBUG`, the
`RCTLogConvertError` gets printed and `nil` is returned, whereas if not
`DEBUG`, the `json` object is returned **unchanged**.
With this PR, the `NSNull` get clean up.

## How I tested this

I used commit
[4b369604d73564159c252483c7bf0801769eecdd](https://github.com/RevenueCat/react-native-purchases/commit/4b369604d73564159c252483c7bf0801769eecdd)
to test this, as the purchasetester app in that commit already uses RN
v0.78.0. With this workaround, the app crashes on iOS right after
launch.