Add default log handler (#1271)
iOS and Android logs are not logged into the JS console. 

This PR sets the native SDK's `logHandler` to a default one that logs
every message coming from the native SDKs to the JS console. This way,
by default, all logs generated in the native SDKs are now visible in the
JS console.

Note that developers can still override the log handler by calling the
existing public API `setLogHandler`.