Add web target and expo test app (#1307)
This adds a web target to our RN SDK and just provides fake responses to
any methods call in a web environment as previously added only for an
Expo Go environment. In order to provide this functionality, this adds
[react-native-web](https://necolas.github.io/react-native-web/) as a
peer dependency, which seems to be the standard way to add a web target
to react native projects and is also used by Expo. This dependency
implements some (but not all) of the RN types but on web.

This also adds a new Expo purchaseTester app, since it's becoming more
useful to be able to test on it. Proper web implementation will be added
future PRs, started in #1330

---------

Co-authored-by: Antonio Pallares <ajpallares@users.noreply.github.com>