Remove `react-native-get-random-values` dependency temporarily (#1364)
This dependency was added to add minimal support for Expo Go. When
creating anonymous user ids, it's not able to find the `crypto` library
in Expo Go, which is used by `purchases-js`.
`react-native-get-random-values` adds that dependency so it works on
Expo Go, however, some devs have reported issues in #1356. While we
haven't been able to repro as of now, there are alternatives we can
follow once we add more functional support to Expo Go.

This PR reverts adding that library for now until we work on those.
React native web support should still work fine on a browser but running
on Expo Go will crash with a `[ReferenceError: Property 'crypto' doesn't
exist]` error. For users trying to make it work, they can potentially
add `react-native-get-random-values` to their dependencies and import
the library somewhere in their app, like: `import
'react-native-get-random-values';`.