Fix compilation in 0.81 (#1400)
`getCurrentActivity` was deprecated in 0.80 (see
https://github.com/facebook/react-native/blob/29c1f72801ecd2dc25b0f2b03c049671f6ba6225/packages/react-native/ReactAndroid/src/main/java/com/facebook/react/bridge/ReactContextBaseJavaModule.kt#L31)

But it looks like they have migrated it to Kotlin (it was in Java
before) in 0.81. Before `getCurrentActivity` (when in Java) would
generate a `currentActivity` accessor. But in Kotlin
`getCurrentActivity` doesn't generate `currentActivity` which breaks
compilation.

https://github.com/f-23/react-native-passkey/issues/73

https://github.com/react-native-datetimepicker/datetimepicker/issues/1002
https://github.com/segmentio/analytics-react-native/issues/1096

We got our own report in
https://github.com/RevenueCat/react-native-purchases/issues/1396

Apart from that it looks like `getDynamic.asMap()` returns optional now