Fix web targets and expo go (#1439)
It seems we broke web and expo go targets with #1408 

The reason is that we added functionality for the log handler but was
not working correctly. Specifically, the plugin was not passing the
handler itself to the browser module method. This is because it differs
from how it works between mobile and web:
- In iOS/Android, the handler is not directly passed to the native side.
Instead, we cache it on the typescript side, and just add an event from
the native side to send events to typescript with new logs, which use
the cached handler to log.
- In web, we can actually pass the given handler to the web SDK, instead
of having to cache it on the typescript side and creating this
indirection with the events.

The problem was that in web (also expo go), we were not passing this
handler to the web SDK, causing an issue while configuring the SDK. This
fixes that.