Removed the async keyword from setupPurchases since no async calls are being made (#1451)
While working on unrelated unit tests I noticed that exceptions thrown
inside `setupPurchases` were not being caught by Jest using
`.expect().toThrow()`. It turns out this is because `setupPurchases()`
is marked as `async`, even though it's not making any `async` calls and
it's not being awaited in
[purchases.ts](https://github.com/RevenueCat/react-native-purchases/blob/main/src/purchases.ts#L320).

Since this is not a public API this change should not have any other
side effects.